requirements.txt
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.7
scikit-learn==1.7.2
Pillow==12.0.0
opencv-python==4.12.0.88
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
matplotlib-inline==0.2.1
tkinterdnd2==0.4.3